Rename redux to store
