+++ src/navigators/AppRouter.tsx
@@ -1,7 +1,7 @@
 import {useAsyncStorage} from '@react-native-async-storage/async-storage';
 import React, {useEffect} from 'react';
 import {useDispatch, useSelector} from 'react-redux';
-import {addAuth, authSelector} from '../redux/reducers/authReducer';
+import {addAuth, authSelector} from '../store/reducers/authReducer';
 import AuthNavigator from './AuthNavigator';
 import MainNavigator from './MainNavigator';
 
@@ -20,7 +20,7 @@ const AppRouter = () => {
   const checkLogin = async () => {
     const res = await getItem();
     res && dispatch(addAuth(JSON.parse(res)));
-		console.log(res);
+    console.log(res);
   };
 
   return <>{auth.accessToken ? <MainNavigator /> : <AuthNavigator />}</>;
