+++ src/screens/home/HomeScreen.tsx
@@ -2,7 +2,7 @@ import AsyncStorage from '@react-native-async-storage/async-storage';
 import React from 'react';
 import {Button, Text, View} from 'react-native';
 import {useDispatch} from 'react-redux';
-import {removeAuth} from '../../redux/reducers/authReducer';
+import {removeAuth} from '../../store/reducers/authReducer';
 import {homeStyles} from './styles';
 
 const HomeScreen = () => {
